Add files via upload
Created a simple website where payment gateway is integrated. There is a simple donate button on homepage, on clicking the donate button, the user will land on the payment page where user can select the amount to be paid and the payment type that includes Credit Card, Netbanking, Paypal and Debit Card. On successful payment an invoice will be generated on a new page regarding completion of the payment.